+++ roxygen2.R
@@ -8,26 +8,26 @@ library(roxygen2)
 ##
 ## (2) Create documentation file(s)
 ##
-document("../meta")
+document()
 
 
 ##
 ## (3) Build R package and PDF file with help pages
 ##
-build("../meta")
-build_manual("../meta")
+build()
+build_manual()
 
 
 ##
 ## (4) Install R package
 ##
-install("../meta")
+install()
 
 
 ##
 ## (5) Check R package
 ##
-check("../meta")
+check()
 
 
 ##
@@ -37,10 +37,3 @@ setwd("..")
 library(numDeriv)
 run_examples("meta", run_dontrun = TRUE, run_donttest = TRUE)
 warnings()
-
-
-##
-## (6) Check R package (with dontrun examples)
-##
-##setwd("meta")
-##check("../meta", run_dont_test = TRUE)